extract future of handle response to a new class
