+++ gateway-ha/src/main/java/io/trino/gateway/proxyserver/ProxyResponse.java
@@ -1,0 +1,30 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.gateway.proxyserver;
+
+import com.google.common.collect.ListMultimap;
+import io.airlift.http.client.HeaderName;
+
+import static java.util.Objects.requireNonNull;
+
+public record ProxyResponse(
+        int statusCode,
+        ListMultimap<HeaderName, String> headers,
+        String body)
+{
+    public ProxyResponse
+    {
+        requireNonNull(headers, "headers is null");
+    }
+}
